backend/requirements.txt
# ===========================================================
# Backend Dependencies for Learning Bot
# ===========================================================
# Install with: pip install -r requirements.txt
# ===========================================================

# -----------------
# Web Framework
# -----------------
fastapi==0.109.0          # Modern async web framework
uvicorn[standard]==0.27.0 # ASGI server to run FastAPI

# -----------------
# Database
# -----------------
sqlalchemy==2.0.25        # ORM for database operations
aiosqlite==0.19.0         # Async SQLite support

# -----------------
# Authentication
# -----------------
python-jose[cryptography]==3.3.0  # JWT token handling
passlib[bcrypt]==1.7.4            # Password hashing
bcrypt==4.0.1                     # Pin bcrypt for passlib compatibility
python-multipart==0.0.6           # Form data parsing

# -----------------
# LLM Integration
# -----------------
mistralai          # Mistral AI API client
httpx            # Async HTTP client for API calls

# -----------------
# RAG (Retrieval Augmented Generation)
# -----------------
chromadb==0.4.22          # Vector database for document storage
sentence-transformers==2.3.1  # Text embeddings
langchain==0.1.4          # LLM orchestration framework
langchain-community==0.0.14   # Community integrations

# -----------------
# Document Processing
# -----------------
pypdf==3.17.4             # PDF text extraction
python-docx==1.1.0        # Word document processing
ebooklib==0.18            # EPUB file processing

# -----------------
# Voice Processing
# -----------------
# Note: Voice-to-text handled on frontend via Web Speech API
# Text-to-speech uses browser's native capabilities
# No additional backend dependencies needed for basic voice

# -----------------
# Utilities
# -----------------
python-dotenv==1.0.0      # Environment variable loading
pydantic==2.5.3           # Data validation
pydantic-settings==2.1.0  # Settings management
pydantic[email]           # Email validation
# -----------------
# Development
# -----------------
pytest==7.4.4             # Testing framework
pytest-asyncio==0.23.3    # Async test support
